Restructurization and first version added
