+++ .gitignore
@@ -3,3 +3,5 @@ dist
 .spec
 .DS_Store
 .vscode
+/__pycache__
+*__pycache__
