+++ requirements.txt
@@ -1,0 +1,2 @@
+dearpygui>=1.8.0
+matplotlib>=3.5.3
