+++ setup.py
@@ -1,0 +1,11 @@
+from setuptools import setup
+
+setup(     
+    name="ROC'n'ROLL",
+    author='SoftwareEngineeringDreamTeam',     
+    version='0.1.0',     
+    install_requires=[         
+        'dearpygui'
+        'matplotlib',
+    ],
+)
